# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ai-marketing-os",
    "description": "AI Marketing Intelligence OS \u2014 authoritative API contract. Frontend API types are generated from this document (openapi-typescript).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List Products",
        "description": "Returns all products in the portfolio with brain summary.",
        "operationId": "list_products_api_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create Product",
        "description": "Adds a new product to the portfolio dynamically.\nNo new worker or crawler code required.",
        "operationId": "create_product_api_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/products/{product_id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Product Detail",
        "description": "Returns full details including complete ProductBrain.",
        "operationId": "get_product_detail_api_products__product_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions": {
      "get": {
        "tags": [
          "Missions"
        ],
        "summary": "List Missions",
        "operationId": "list_missions_api_missions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "Create Mission",
        "operationId": "create_mission_api_missions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MissionCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Name"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "normal"
          },
          "budget_usd": {
            "type": "number",
            "title": "Budget Usd",
            "default": 100.0
          },
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Ids"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/MissionGoalIn"
            },
            "type": "array",
            "title": "Goals"
          }
        },
        "type": "object",
        "required": [
          "name",
          "objective"
        ],
        "title": "MissionCreateRequest"
      },
      "MissionGoalIn": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "MissionGoalIn"
      },
...
      "ProductCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Slug"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "SaaS"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "icp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icp",
            "default": ""
          },
          "problems_solved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems Solved"
          },
          "problems_not_solved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Problems Not Solved"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "description"
        ],
        "title": "ProductCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}